symbols and keywords' token added
